feature/wheel: add private weapon and armor equipment favorites

+++ src/WheelView.h
@@ -3,7 +3,7 @@
 #include <imgui.h>
 namespace wheel {
 struct View { float animation{}; };
-struct Action { std::uint32_t hoveredItem{}; bool configHovered{}; };
+struct Action { std::uint64_t hoveredItem{}; bool configHovered{}; };
 Action drawWheel(Model& model, View& view, ImVec2 position={}, ImVec2 size={});
 Model demoInventory();
 }
